Move StripeCreditCardFields and CreditCardFields into lib/payment-methods
